AKPW-11 Refactor GlobalViewModel.kt and create segregated
