+++ app/src/main/java/com/creativegrpcx/perawatcher/dependencyInjector/ViewModelModule.kt
@@ -15,4 +15,5 @@ class ViewModelModule {
     fun providesGlobalViewModelFactory (application: Application,
                                         userRepository: DataRepository
     ) = GlobalViewModelFactory(application,userRepository)
+
 }
