+++ app/src/main/java/com/creativegrpcx/perawatcher/domain/model/state/AddWalletState.kt
@@ -1,4 +1,4 @@
-package com.creativegrpcx.perawatcher.domain.model
+package com.creativegrpcx.perawatcher.domain.model.state
 
 import com.creativegrpcx.perawatcher.domain.types.WalletType
 
@@ -7,3 +7,5 @@ data class AddWalletState(
     val walletType: WalletType = WalletType.CASH,
     val walletAmount : String = "",
 )
+
+
