+++ app/src/main/java/com/creativegrpcx/perawatcher/domain/model/state/ErrorState.kt
@@ -1,4 +1,4 @@
-package com.creativegrpcx.perawatcher.domain.model
+package com.creativegrpcx.perawatcher.domain.model.state
 
 data class ErrorState(
     val isShowed : Boolean = false,
