+++ app/src/main/java/com/creativegrpcx/perawatcher/domain/viewmodel/GlobalViewModelFactory.kt
@@ -11,14 +11,13 @@ class GlobalViewModelFactory @Inject constructor(
     private val repository: DataRepository
 ) : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-//        if (modelClass.isAssignableFrom(GlobalViewModel::class.java)) {
-//            @Suppress("UNCHECKED_CAST")
-//            return GlobalViewModel(application,repository) as T
-//        }
-//
         @Suppress("UNCHECKED_CAST")
         return when (modelClass) {
-            GlobalViewModel::class.java -> GlobalViewModel(application, repository) as T
+            DashboardViewModel::class.java -> DashboardViewModel(application,repository) as T
+            HistoryViewModel::class.java -> HistoryViewModel(application,repository) as T
+            WalletViewModel::class.java -> WalletViewModel(application,repository) as T
+            AddTransactionViewModel::class.java -> AddTransactionViewModel(application, repository) as T
+            AddWalletViewModel::class.java -> AddWalletViewModel(application, repository) as T
             else -> {
                 throw IllegalArgumentException("Unknown ViewModel class")
             }
