+++ app/src/main/java/com/creativegrpcx/perawatcher/ui/activity/ComposeActivity.kt
@@ -5,25 +5,18 @@ import android.os.Bundle
 import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.activity.viewModels
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
 import com.creativegrpcx.perawatcher.MainApplication
-import com.creativegrpcx.perawatcher.domain.viewmodel.GlobalViewModel
 import com.creativegrpcx.perawatcher.domain.viewmodel.GlobalViewModelFactory
 import com.creativegrpcx.perawatcher.ui.components.ApplicationHeader
 import com.creativegrpcx.perawatcher.ui.nav.ApplicationRoute
 import com.creativegrpcx.perawatcher.ui.theme.AppTheme
 import com.google.accompanist.navigation.animation.rememberAnimatedNavController
-import kotlinx.coroutines.flow.collectLatest
-import java.util.*
 import javax.inject.Inject
 
 class ComposeActivity : ComponentActivity() {
@@ -34,14 +27,12 @@ class ComposeActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         (application as MainApplication).component.inject(this)
         super.onCreate(savedInstanceState)
-        val globalViewModel: GlobalViewModel by viewModels {
-            globalViewModelFactory
-        }
+
         try {
             setContent {
                 AppTheme {
                     MainScreen(
-                        globalViewModel
+                        globalViewModelFactory
                     )
                 }
             }
@@ -55,52 +46,29 @@ class ComposeActivity : ComponentActivity() {
 @OptIn(ExperimentalMaterial3Api::class, ExperimentalAnimationApi::class)
 @Composable
 fun MainScreen(
-    globalViewModel: GlobalViewModel,
+    factory: GlobalViewModelFactory,
 ) {
     val navigationController = rememberAnimatedNavController()
-    val state = globalViewModel.routeState.collectAsState().value
     val snackbarHostState = remember {
         SnackbarHostState()
     }
 
     // A surface container using the 'background' color from the theme
 
-    if (state.isRouteChange) navigationController.navigate(state.currentRoute.route) {
-        launchSingleTop = true
-        if (state.oldRoute != null){
-            popUpTo(state.oldRoute.route)
-        }
-    }
-    if (state.isNavigateUp) navigationController.navigateUp()
-
-
-    LaunchedEffect(Unit) {
-        globalViewModel.errorState.collectLatest {
-            if (it.isShowed) {
-                snackbarHostState.showSnackbar(
-                    message = it.message,
-                    duration = SnackbarDuration.Short
-                )
-            }
-        }
-    }
-
     Scaffold(
         modifier = Modifier.fillMaxSize(),
         containerColor = MaterialTheme.colorScheme.background,
         topBar = {
             ApplicationHeader(
-                newRoute = state.currentRoute,
-            ) {
-                globalViewModel.updateCurrentRoute(it)
-            }
+                navigation = navigationController,
+            )
         },
         snackbarHost = {
             SnackbarHost(hostState = snackbarHostState)
         }
     ) {
         ApplicationRoute(
-            viewModel = globalViewModel,
+            factory = factory,
             padding = it,
             navHostController = navigationController
         )
