+++ app/src/main/java/com/creativegrpcx/perawatcher/ui/components/ApplicationHeader.kt
@@ -6,23 +6,27 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.AccountCircle
 import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import androidx.navigation.compose.currentBackStackEntryAsState
 import com.creativegrpcx.perawatcher.ui.nav.NavigationRoute
-import com.creativegrpcx.perawatcher.ui.nav.ScreenRoute
 import com.creativegrpcx.perawatcher.ui.utils.Constants
+import com.creativegrpcx.perawatcher.ui.utils.currentRoute
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun ApplicationHeader(
-    newRoute: ScreenRoute,
-    onRouteChange: (route: ScreenRoute?) -> Unit
+    navigation: NavController,
 ) {
+
+    val navigationState by navigation.currentBackStackEntryAsState()
+    val latestRoute = navigationState?.destination?.route.currentRoute()
+
     Column(
         modifier = Modifier
             .wrapContentHeight()
@@ -31,28 +35,20 @@ fun ApplicationHeader(
         horizontalAlignment = Alignment.Start,
     ) {
 
-        when (newRoute.route) {
+        when (latestRoute.route) {
             NavigationRoute.AddTransaction.withoutArgs.route,
             NavigationRoute.AddWallet.withoutArgs.route -> {
                 IconButton(onClick = {
-                    onRouteChange(
-                        null
-                    )
+                    navigation.navigateUp()
                 }) {
                     Icon(imageVector = Icons.Default.ArrowBack, contentDescription = "")
                 }
             }
-            NavigationRoute.PopUpAddWallet.withoutArgs.route ->{
-
+            NavigationRoute.PopUpAddWallet.withoutArgs.route -> {
             }
             else -> {
                 IconButton(onClick = {
-                    onRouteChange(
-                        ScreenRoute(
-                            name = "profile",
-                            route = "profile"
-                        )
-                    )
+
                 }) {
 
                     Icon(imageVector = Icons.Default.AccountCircle, contentDescription = "")
@@ -68,9 +64,11 @@ fun ApplicationHeader(
                     items(Constants.MainScreenRoutes) { route ->
                         NavigationChipItem(
                             routeContent = route,
-                            currentRoute = newRoute,
+                            currentRoute = latestRoute,
                         ) {
-                            onRouteChange(it)
+                            navigation.navigate(it.route) {
+                                launchSingleTop = true
+                            }
                         }
                     }
                 }
