+++ app/src/main/java/com/creativegrpcx/perawatcher/ui/nav/Navigation.kt
@@ -1,6 +1,5 @@
 package com.creativegrpcx.perawatcher.ui.nav
 
-import androidx.activity.compose.BackHandler
 import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.animation.slideInHorizontally
 import androidx.compose.animation.slideInVertically
@@ -10,8 +9,9 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavHostController
-import com.creativegrpcx.perawatcher.domain.viewmodel.GlobalViewModel
+import com.creativegrpcx.perawatcher.domain.viewmodel.GlobalViewModelFactory
 import com.creativegrpcx.perawatcher.ui.screens.*
 import com.google.accompanist.navigation.animation.AnimatedNavHost
 import com.google.accompanist.navigation.animation.composable
@@ -20,20 +20,24 @@ import com.google.accompanist.navigation.animation.composable
 @OptIn(ExperimentalAnimationApi::class)
 @Composable
 fun ApplicationRoute(
-    viewModel : GlobalViewModel,
-    padding : PaddingValues,
+    factory: GlobalViewModelFactory,
+    padding: PaddingValues,
     navHostController: NavHostController
-){
+) {
 
     AnimatedNavHost(
         modifier = Modifier
-            .padding(PaddingValues(
-                start = 16.dp,
-                top = padding.calculateTopPadding(),
-                end = 16.dp))
+            .padding(
+                PaddingValues(
+                    start = 16.dp,
+                    top = padding.calculateTopPadding(),
+                    end = 16.dp
+                )
+            )
             .fillMaxSize(),
         navController = navHostController,
-        startDestination = NavigationRoute.Dashboard.withoutArgs.route){
+        startDestination = NavigationRoute.Dashboard.withoutArgs.route
+    ) {
         composable(
             route = NavigationRoute.Dashboard.withoutArgs.route,
             enterTransition = {
@@ -41,7 +45,12 @@ fun ApplicationRoute(
                     initialOffsetX = { 1000 })
             },
             content = {
-                DashboardScreen(viewModel)
+                DashboardScreen(
+                    viewModel =  viewModel(
+                        factory = factory
+                    ),
+                    navController = navHostController
+                )
             },
         )
         composable(
@@ -60,7 +69,11 @@ fun ApplicationRoute(
                 slideInHorizontally(initialOffsetX = { 1000 })
             },
             content = {
-                HistoryScreen(viewModel)
+                HistoryScreen(
+                    viewModel(
+                        factory = factory
+                    )
+                )
             }
         )
         composable(
@@ -69,19 +82,24 @@ fun ApplicationRoute(
                 slideInHorizontally(initialOffsetX = { 1000 })
             },
             content = {
-                WalletScreen(viewModel)
+                WalletScreen(
+                    navigation = navHostController,
+                    viewModel(
+                        factory = factory
+                    )
+                )
             }
         )
         composable(
             route = NavigationRoute.AddTransaction.withoutArgs.route,
             enterTransition = {
-                slideInVertically (
+                slideInVertically(
                     initialOffsetY = { 1000 }
                 )
             },
             content = {
                 AddTransactionScreen(
-                    viewModel =  viewModel
+                    viewModel = viewModel(factory = factory)
                 )
             }
         )
@@ -89,34 +107,29 @@ fun ApplicationRoute(
         composable(
             route = NavigationRoute.AddWallet.withoutArgs.route,
             enterTransition = {
-                slideInVertically (
+                slideInVertically(
                     initialOffsetY = { 1000 }
                 )
             },
             content = {
                 AddWalletScreen(
-                    viewModel =  viewModel
+                    viewModel = viewModel(factory = factory)
                 )
             }
         )
-
-        composable(
-            route = NavigationRoute.PopUpAddWallet.withoutArgs.route,
-            enterTransition = {
-                slideInVertically (
-                    initialOffsetY = { 1000 }
-                )
-            },
-            content = {
-               PopUpFullScreen (
-                   viewModel =  viewModel,
-                )
-            }
-        )
+//
+//        composable(
+//            route = NavigationRoute.PopUpAddWallet.withoutArgs.route,
+//            enterTransition = {
+//                slideInVertically(
+//                    initialOffsetY = { 1000 }
+//                )
+//            },
+//            content = {
+//                PopUpFullScreen(
+//                    viewModel = viewModel,
+//                )
+//            }
+//        )
     }
-
-    BackHandler {
-        viewModel.updateCurrentRoute(null)
-    }
-
 }
