+++ app/src/main/java/com/creativegrpcx/perawatcher/ui/nav/NavigationRoute.kt
@@ -7,8 +7,8 @@ sealed class NavigationRoute(
     object Statistics : NavigationRoute(ScreenRoute( route = "statistics", name = "statistics"))
     object History : NavigationRoute(ScreenRoute( route = "history", name = "history"))
     object Wallet : NavigationRoute(ScreenRoute( route = "wallet", name = "wallet"))
-    object AddTransaction : NavigationRoute(ScreenRoute( route = "add_trans", name = "add_trans"))
-    object AddWallet : NavigationRoute(ScreenRoute( route = "add_wallet", name = "add_wallet"))
+    object AddTransaction : NavigationRoute(ScreenRoute( route = "add_trans", name = "add_trans" , isPopInclusive = true , popUpTo = "dashboard"))
+    object AddWallet : NavigationRoute(ScreenRoute( route = "add_wallet", name = "add_wallet", isPopInclusive = true , popUpTo = "wallet"))
     object PopUpAddWallet : NavigationRoute(ScreenRoute( route = "pop_up_add_wallet", name = "pop_up_add_wallet", isPopUp = true))
 
     fun withArgs(vararg args : Any): ScreenRoute{
@@ -26,5 +26,7 @@ data class ScreenRoute(
     val name : String,
     val args : List<Any>? = null,
     val isSelected : Boolean = false,
-    val isPopUp : Boolean = false
+    val isPopUp : Boolean = false,
+    val isPopInclusive : Boolean = false,
+    val popUpTo : String = ""
 )
